+++ baseapp/baseapp.go
@@ -24,6 +24,7 @@ import (
 
 	"cosmossdk.io/log/v2"
 	"cosmossdk.io/store"
+	"cosmossdk.io/store/rootmulti"
 	"cosmossdk.io/store/snapshots"
 	storetypes "cosmossdk.io/store/types"
 
@@ -236,7 +237,11 @@ func NewBaseApp(
 		app.SetVerifyVoteExtensionHandler(NoOpVerifyVoteExtensionHandler())
 	}
 	if app.interBlockCache != nil {
-		app.cms.SetInterBlockCache(app.interBlockCache)
+		if rms, ok := app.cms.(*rootmulti.Store); ok {
+			rms.SetInterBlockCache(app.interBlockCache)
+		} else {
+			logger.Warn("SetInterBlockCache: CommitMultiStore is not rootmulti.Store, option ignored")
+		}
 	}
 
 	app.runTxRecoveryMiddleware = newDefaultRecoveryMiddleware()